requirements.txt
django 
djangorestframework 
django-cors-headers 
djangorestframework-simplejwt 
psycopg2-binary 
paypalrestsdk 
openai 
#smtplib
python-dotenv